+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## TODO
 
+### Changed
+
+- The game class generator now generates the game constructor and the static `Construct` method.
+
 ### Fixed
 
 - Fixed not always generating required types.